added new user setup,searching

+++ index.js
@@ -57,20 +57,27 @@ program
     }
   });
 
-(function isInitialised(CONFIG) {
-  if (Object.keys(CONFIG).length === 0) {
-    console.log(chalk.red.bold('notes directory not set !!\n'));
+function isInitialised(CONFIG) {
+  if (CONFIG.INIT === false) {
+    // console.log(chalk.red.bold('notes directory not set !!\n'));
     console.log(chalk.blue("use 'gitn config -g <absolute_path_to_notes_directory>' to initialise gitn"));
-    return;
+    return false;
   }
-}(CONFIG));
+  return true;
+}
+
 
 
 program
-  .command('status ')
+  .command('status')
   .description('outputs current branch')
   .action(() => {
-    console.log(chalk.green(`On branch ${chalk.bold.blue(cfg.get('CURRENT_BRANCH'))}\n`));
+    try {
+      if (!isInitialised(CONFIG)) throw new Error("Initialize notes directory to begin !");
+      console.log(chalk.green(`On branch ${chalk.bold.blue(cfg.get('CURRENT_BRANCH'))}\n`));
+    } catch (e) {
+      console.log(chalk.red(e.message));
+    }
   })
 
 program
@@ -80,21 +87,44 @@ program
   .action(async (args) => {
     let filehandle;
     try {
+      if (!isInitialised(CONFIG)) throw new Error("Initialize notes directory to begin !");
       filehandle = await fs.open(`${NOTE_DIR}${CURRENT_BRANCH}.txt`, 'r');
       data = await filehandle.readFile('utf-8');
       console.log(chalk.green(`On branch ${chalk.bold.blue(CURRENT_BRANCH)}`));
       notes = data.split("\n");
       notes.pop();
+      if(notes.length===0){
+        console.log(chalk.red('no notes found'));
+        return;
+      }
       if (args.grep) {
 
         let re = new RegExp(args.grep);
-        notes = notes.filter(e => {
-          return re.test(e)
+
+        notes = notes.map(e => {
+          return e.split(" ");
+        })
+
+        for (let i = 0; i < notes.length; i++) {
+          let noteToPrint = "";
+          for (let j = 0; j < notes[i].length; j++) {
+            if (notes[i][j].trim().length != 0 && re.test(notes[i][j])) {
+              noteToPrint += `${chalk.blue(notes[i][j])} `;
+            } else {
+              noteToPrint += `${notes[i][j]} `;
+            }
+          }
+          console.log(`${chalk.bold.blue('├─ ')}${noteToPrint}`);
+        }
+
+      } else {
+        notes.forEach(e => {
+          console.log(`${chalk.bold.blue('├─ ')}${e}`)
         })
       }
-      notes.forEach(e => {
-        console.log(`${chalk.bold.blue('├─ ')}${e}`)
-      })
+
+    } catch (e) {
+      console.log(chalk.red(e));
     } finally {
       if (filehandle !== undefined)
         await filehandle.close();
@@ -108,6 +138,7 @@ program
   .action(async (args) => {
     let filehandle;
     try {
+      if (!isInitialised(CONFIG)) throw new Error("Initialize notes directory to begin !");
       filehandle = await fs.open(`${NOTE_DIR}${CURRENT_BRANCH}.txt`, 'a', 666);
       await fs.appendFile(filehandle, args.message + '\n', 'utf-8');
     } catch (e) {
@@ -127,6 +158,15 @@ program
   .action(async (branch) => {
     let files;
     let filehandle;
+
+    try {
+      if (!isInitialised(CONFIG)) throw new Error("Initialize notes directory to begin !");
+    } catch (e) {
+      console.log(chalk.red(e.message));
+      return;
+    }
+
+
     if (program.args.includes('-d') || program.args.includes('--delete')) {
       try {
         if (!branch) throw new Error(" option '-d --delete' requires <branch> to be specified");
@@ -145,13 +185,13 @@ program
         files = await fs.readdir(`${NOTE_DIR}`, 'utf-8');
         for (let i = 0; i < files.length; i++) {
           files[i] = files[i].split('.')[0]
-          branchNotesTree[files[i]] = {};
+          branchNotesTree[`${chalk.bold.blue(files[i])}`] = {};
           filehandle = await fs.open(`${NOTE_DIR}${files[i]}.txt`, 'r');
           data = await filehandle.readFile('utf-8');
           notes = data.split("\n");
           notes.pop();
           notes.forEach(e => {
-            branchNotesTree[files[i]][e] = null;
+            branchNotesTree[`${chalk.bold.blue(files[i])}`][`${chalk.green(e)}`] = null;
           })
         }
         console.log(treeify.asTree(branchNotesTree, true));
@@ -176,7 +216,9 @@ program
       }
     } else {
       try {
-        filehandle = await fs.open(`${NOTE_DIR}${branch}.txt`, 'wx'); // wx fails if exist, so may use w 
+        filehandle = await fs.open(`${NOTE_DIR}${branch}.txt`, 'wx'); // wx 
+        console.log(chalk.green(`new branch ${chalk.bold.blue(branch)} created.`))
+        //fails if exist, so may use w 
         // console.log(filehandle)
       } catch (e) {
         if (e.code == 'EEXIST') {
@@ -197,10 +239,14 @@ program
   .description('deletes all commits on current branch')
   .action(async (args) => {
     try {
+
+      if (!isInitialised(CONFIG)) throw new Error("Initialize notes directory to begin !");
+
       await fs.truncate(`${NOTE_DIR}${CURRENT_BRANCH}.txt`);
       console.log(`${chalk.green('reset successful')}`)
     } catch (e) {
-      console.log(e.message)
+      console.log(e.message);
+      return;
     }
 
   })
@@ -210,6 +256,14 @@ program
   .option('-b --newBranch', 'create a new branch')
   .description('create/switch branch')
   .action(async (branch) => {
+
+    try {
+      if (!isInitialised(CONFIG)) throw new Error("Initialize notes directory to begin !");
+    } catch (e) {
+      console.log(chalk.red(e.message));
+      return;
+    }
+
     let filehandle;
     if (program.args.includes('-b') || program.args.includes('--createBranch')) {
       try {
@@ -257,7 +311,10 @@ program
     let filehandle;
     let currentFilehandle;
     try {
-      if (branch === CURRENT_BRANCH) throw ({ errno: 8989, msg: 'on same branch' })
+
+      if (!isInitialised(CONFIG)) throw new Error("Initialize notes directory to begin !");
+
+      if (branch === CURRENT_BRANCH) throw ({ errno: 8989, message: 'on same branch' })
       filehandle = await fs.open(`${NOTE_DIR}${branch}.txt`, 'r');
       data = await filehandle.readFile('utf-8');
       currentFilehandle = await fs.open(`${NOTE_DIR}${CURRENT_BRANCH}.txt`, 'a', 666);
@@ -267,9 +324,9 @@ program
       if (e.errno === -2) {
         console.log(chalk.red('no such branch'));
       } else if (e.errno === 8989) {
-        console.log(chalk.red(e.msg));
+        console.log(chalk.red(e.message));
       } else
-        console.log(e);
+        console.log(e.message);
     } finally {
       if (filehandle !== undefined)
         await filehandle.close();
@@ -285,6 +342,8 @@ program
   .action(async () => {
     let filehandle;
     try {
+
+      if (!isInitialised(CONFIG)) throw new Error("Initialize notes directory to begin !");
       filehandle = await fs.open(`${NOTE_DIR}${CURRENT_BRANCH}.txt`, "r+");
       data = await filehandle.readFile('utf-8');
       if (filehandle !== undefined)
@@ -319,7 +378,7 @@ program
       }
 
     } catch (e) {
-      console.log(e)
+      console.log(chalk.red(e.message));
     }
 
   })
